Fix previous fix. Subtracting a uint from a uint gives a uint, so cast to int first.
svn path=/trunk/koffice/; revision=1067824

+++ krita/plugins/filters/blur/kis_gaussian_blur_filter.cpp
@@ -97,7 +97,7 @@ void KisGaussianBlurFilter::process(KisConstProcessingInformation srcInfo,
 
     for (uint x = 0; x < horizKernelSize; x++)
     {
-        uint xDistance = qAbs(horizontalRadius - x);
+        uint xDistance = qAbs((int)horizontalRadius - (int)x);
         horizGaussian(0, x) = horizMultiplicand * exp( -(qreal)((xDistance * xDistance) + (horizontalRadius * horizontalRadius)) * horizExponentMultiplicand );
     }
 
@@ -111,7 +111,7 @@ void KisGaussianBlurFilter::process(KisConstProcessingInformation srcInfo,
 
     for (uint y = 0; y < verticalKernelSize; y++)
     {
-        uint yDistance = qAbs(verticalRadius - y);
+        uint yDistance = qAbs((int)verticalRadius - (int)y);
         verticalGaussian(y, 0) = verticalMultiplicand * exp( -(qreal)((yDistance * yDistance) + (verticalRadius * verticalRadius)) * verticalExponentMultiplicand );
     }
